live refactoring: fixing packages
