+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/ActionCluster.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql
+
 class ActionCluster {
 
 // create the structure properly later
