+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/Table.kt
@@ -1,3 +1,6 @@
+package org.evomaster.core.problem.graphql
+
+
 class Table {
     var tableField : String?=null
     var tableType : String?=null
