+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/Data.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 class Data {
 
     var __schema: __Schema? = null
