+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/SchemaObj.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 class SchemaObj {
 
     var data: Data? = null
