+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/TypeRef.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 class TypeRef {
 
      var kind: __TypeKind?=null
