+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/__Directive.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 class __Directive {
 
      var name: String?=null
