+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/__DirectiveLocation.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 enum class __DirectiveLocation {
