+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/__Type.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 class __Type {
 
      var name: String?= null
