+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/__TypeKind.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 enum class __TypeKind {
 
         SCALAR, OBJECT, INTERFACE, UNION, ENUM, INPUT_OBJECT, LIST, NON_NULL
