+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/schema/ofTypeOn__Type.kt
@@ -1,3 +1,5 @@
+package org.evomaster.core.problem.graphql.schema
+
 class ofTypeOn__Type {
 
          var name: String?= null
