+++ core/src/test/resources/graphql/QueryTypeGlobalPetsClinic.json
